@@ -13,33 +13,46 @@
   "dependencies": {
     "@clockwork-xyz/sdk": "^0.3.0",
     "@coral-xyz/anchor": "^0.27.0",
-    "@docusaurus/core": "2.4.0",
-    "@docusaurus/preset-classic": "2.4.0",
-    "@docusaurus/theme-mermaid": "2.4.0",
+    "@docusaurus/core": "^3.1.1",
+    "@docusaurus/plugin-google-gtag": "^3.1.1",
+    "@docusaurus/preset-classic": "^3.1.1",
+    "@docusaurus/theme-mermaid": "^3.1.1",
+    "@helium/account-fetch-cache": "^0.6.29",
+    "@helium/account-fetch-cache-hooks": "^0.6.29",
     "@helium/address": "^4.8.1",
-    "@helium/fanout-sdk": "^0.1.1",
-    "@helium/helium-react-hooks": "^0.1.1",
-    "@helium/idls": "^0.1.1",
-    "@helium/spl-utils": "^0.1.1",
-    "@mdx-js/react": "^1.5.8",
+    "@helium/data-credits-sdk": "^0.6.29",
+    "@helium/fanout-sdk": "^0.6.29",
+    "@helium/helium-react-hooks": "^0.6.29",
+    "@helium/idls": "^0.6.29",
+    "@helium/spl-utils": "^0.6.29",
+    "@mdx-js/react": "^3.0.0",
     "@metaplex-foundation/mpl-token-metadata": "^2.8.4",
+    "@pythnetwork/client": "^2.19.0",
     "@solana/spl-account-compression": "^0.1.7",
     "@solana/spl-token": "^0.3.7",
     "@solana/wallet-adapter-ledger": "^0.9.24",
+    "@solana/wallet-adapter-phantom": "^0.9.23",
     "@solana/wallet-adapter-react": "^0.15.32",
     "@solana/wallet-adapter-react-ui": "^0.9.31",
+    "@solana/wallet-adapter-solflare": "^0.6.24",
     "@solana/web3.js": "^1.66.1",
     "axios": "^1.3.4",
     "clsx": "^1.1.1",
+    "docslab-docusaurus": "^0.2.7",
+    "fs": "^0.0.1-security",
     "hast-util-is-element": "1.1.0",
+    "mdx-embed": "^1.1.2",
+    "prism-react-renderer": "^2.3.1",
     "process": "^0.11.10",
-    "react": "^16.8.4",
+    "react": "^18.2.0",
     "react-async-hook": "^4.0.0",
-    "react-dom": "^16.8.4",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
+    "react-table": "^7.8.0",
     "rehype-katex": "5",
     "remark-math": "3",
-    "yarn": "^1.22.10"
+    "util": "^0.12.5",
+    "yarn": "^1.22.21"
   },
   "browserslist": {
     "production": [
@@ -52,5 +65,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "engines": {
+    "node": ">=18.0"
   }
 }
