@@ -1,8 +1,13 @@
 import { ClockworkProvider } from '@clockwork-xyz/sdk'
-import { AnchorProvider, BN } from '@coral-xyz/anchor'
-import { fanoutKey, init, membershipCollectionKey, membershipVoucherKey } from '@helium/fanout-sdk'
-import { AccountProvider, useIdlAccount, useTokenAccount } from '@helium/helium-react-hooks'
-import { IDL } from '@helium/idls/fanout'
+import { AnchorProvider, BN, Program } from '@coral-xyz/anchor'
+import {
+  PROGRAM_ID,
+  fanoutKey,
+  init,
+  membershipCollectionKey,
+  membershipVoucherKey,
+} from '@helium/fanout-sdk'
+import { useIdlAccount, useTokenAccount } from '@helium/helium-react-hooks'
 import { Fanout } from '@helium/idls/lib/types/fanout'
 import { HNT_MINT, searchAssets, sendInstructions, toNumber } from '@helium/spl-utils'
 import {
@@ -10,10 +15,7 @@ import {
   createInitializeMintInstruction,
   getAssociatedTokenAddressSync,
 } from '@solana/spl-token'
-import {
-  useConnection,
-  useWallet
-} from '@solana/wallet-adapter-react'
+import { useConnection, useWallet } from '@solana/wallet-adapter-react'
 import { Keypair, LAMPORTS_PER_SOL, PublicKey, SystemProgram } from '@solana/web3.js'
 import React, { useEffect, useMemo } from 'react'
 import { useAsync, useAsyncCallback } from 'react-async-hook'
@@ -26,7 +28,7 @@ import { Wallet, WrapWithAccountProvider } from './components/Wallet'
 // Default styles that can be overridden by your app
 require('@solana/wallet-adapter-react-ui/styles.css')
 
-export const HstManagerImpl = () => {
+export const HstManagerImpl = ({ idl }: { idl: Fanout }) => {
   const { publicKey, wallet } = useWallet()
   const fanoutK = useMemo(() => fanoutKey('HST')[0], [])
   const collectionAddress = useMemo(() => membershipCollectionKey(fanoutK)[0], [fanoutK])
@@ -38,7 +40,9 @@ export const HstManagerImpl = () => {
         ownerAddress: publicKey?.toBase58(),
         creatorAddress: null,
       })
-      return assets[0]
+      if (assets) {
+        return assets[0]
+      }
     }
 
     return undefined
@@ -60,8 +64,7 @@ export const HstManagerImpl = () => {
       return await init(new AnchorProvider(connection, wallet.adapter, { commitment: 'confirmed' }))
     }
   }, [wallet, connection])
-
-  const { info: fanout } = useIdlAccount<Fanout>(fanoutK, IDL as Fanout, 'fanoutV0')
+  const { info: fanout } = useIdlAccount<Fanout>(fanoutK, idl as Fanout, 'FanoutV0')
   const hst = useMemo(() => fanout && fanout.membershipMint, [fanout])
   const stakeAccountKey = useMemo(
     () => voucher && hst && getAssociatedTokenAddressSync(hst, voucher, true),
@@ -214,11 +217,25 @@ function threadKey(authority: PublicKey, threadId: string): [PublicKey, number] 
   )
 }
 
+export const WrapWithIdl = () => {
+  const { wallet } = useWallet()
+  const { connection } = useConnection()
+  const { result: idl } = useAsync(async () => {
+    return await Program.fetchIdl(PROGRAM_ID, new AnchorProvider(connection, wallet.adapter, {}))
+  }, [])
+
+  if (idl) {
+    return <HstManagerImpl idl={idl as Fanout} />
+  }
+
+  return <div />
+}
+
 export const HstManager = () => {
   return (
     <Wallet>
       <WrapWithAccountProvider>
-        <HstManagerImpl />
+        <WrapWithIdl />
       </WrapWithAccountProvider>
     </Wallet>
   )
